test: split stop admission and stop-gate coverage by invariant
Move looperd stop/admission lifecycle cases and API retry/start/reuse
stop-gate races out of oversized subsystem test files into focused
*_test.go files so each PR addition stays reviewable.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
