+++ cmd/looperd/main_test.go
@@ -477,232 +477,6 @@ func TestCloseLoopTerminatesLoopAndCancelsActiveQueue(t *testing.T) {
 		t.Fatalf("Queue.GetByID() = %#v, want cancelled queue item with close reason", storedQueue)
 	}
 }
-
-// After a successful terminal close, per-loop spawn admission stays closed so
-// in-flight runners cannot AdmitSpawn past the durable terminate.
-func TestCloseLoopKeepsSpawnAdmissionClosedAfterSuccess(t *testing.T) {
-	ctx := context.Background()
-	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	if _, err := coordinator.MigrationRunner().RunPending(ctx); err != nil {
-		t.Fatalf("MigrationRunner().RunPending() error = %v", err)
-	}
-	t.Cleanup(func() { _ = coordinator.Close() })
-
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.April, 21, 12, 0, 0, 0, time.UTC)
-	nowISO := "2026-04-21T12:00:00.000Z"
-	project := storage.ProjectRecord{ID: "project_1", Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Projects.Upsert(ctx, project); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loop := storage.LoopRecord{ID: "loop_close_sticky", Seq: 32, ProjectID: project.ID, Type: "worker", TargetType: "project", Status: "running", CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Loops.Upsert(ctx, loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	services := looperdruntime.Services{
-		Coordinator:      coordinator,
-		Repositories:     repos,
-		Loops:            &loops.Service{DB: coordinator.DB(), Repos: repos, Now: func() time.Time { return now }},
-		ActiveExecutions: registry,
-	}
-
-	if _, err := closeLoop(ctx, services, loop.ID, "Closed by test", func() time.Time { return now }, nil, nil); err != nil {
-		t.Fatalf("closeLoop() error = %v", err)
-	}
-	if !registry.LoopStopActive(loop.ID) {
-		t.Fatal("LoopStopActive = false after successful closeLoop, want sticky closed gate")
-	}
-	_, err = registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: loop.ID, RunID: "run_late", ExecutionID: "exec_late",
-	})
-	if !errors.Is(err, agent.ErrSpawnLoopStopping) {
-		t.Fatalf("AdmitSpawn after closeLoop error = %v, want ErrSpawnLoopStopping", err)
-	}
-}
-
-// After a successful stop, per-loop spawn admission stays closed so an in-flight
-// runner that reaches AgentExecutor.Start after halt returns cannot AdmitSpawn.
-func TestStopLoopKeepsSpawnAdmissionClosedAfterReturn(t *testing.T) {
-	ctx := context.Background()
-	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	if _, err := coordinator.MigrationRunner().RunPending(ctx); err != nil {
-		t.Fatalf("MigrationRunner().RunPending() error = %v", err)
-	}
-	t.Cleanup(func() { _ = coordinator.Close() })
-
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.April, 21, 12, 0, 0, 0, time.UTC)
-	nowISO := "2026-04-21T12:00:00.000Z"
-	project := storage.ProjectRecord{ID: "project_1", Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Projects.Upsert(ctx, project); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loop := storage.LoopRecord{ID: "loop_1", Seq: 30, ProjectID: project.ID, Type: "worker", TargetType: "project", Status: "running", CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Loops.Upsert(ctx, loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	services := looperdruntime.Services{
-		Coordinator:      coordinator,
-		Repositories:     repos,
-		Loops:            &loops.Service{DB: coordinator.DB(), Repos: repos, Now: func() time.Time { return now }},
-		ActiveExecutions: registry,
-	}
-
-	if _, err := stopLoop(ctx, services, loop.ID, "Stopped by test", func() time.Time { return now }, nil, nil); err != nil {
-		t.Fatalf("stopLoop() error = %v", err)
-	}
-	if !registry.LoopStopActive(loop.ID) {
-		t.Fatal("LoopStopActive = false after successful stopLoop, want sticky closed gate")
-	}
-	_, err = registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: loop.ID, RunID: "run_late", ExecutionID: "exec_late",
-	})
-	if !errors.Is(err, agent.ErrSpawnLoopStopping) {
-		t.Fatalf("AdmitSpawn after stopLoop error = %v, want ErrSpawnLoopStopping", err)
-	}
-
-	// Intentional re-activation reopens the gate (unpause/retry/fresh claim path).
-	registry.ClearLoopStop(loop.ID)
-	if _, err := registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: loop.ID, RunID: "run_resume", ExecutionID: "exec_resume",
-	}); err != nil {
-		t.Fatalf("AdmitSpawn after ClearLoopStop error = %v, want success", err)
-	}
-}
-
-// Terminal close must not BeginLoopStop (lease cancel is irreversible for
-// execution.run) before abortable run/execution preflight. A transient lookup
-// error leaves the loop running, so active agents must not be half-killed.
-func TestCloseLoopDoesNotCancelLeasesWhenRunLookupFails(t *testing.T) {
-	ctx := context.Background()
-	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	if _, err := coordinator.MigrationRunner().RunPending(ctx); err != nil {
-		t.Fatalf("MigrationRunner().RunPending() error = %v", err)
-	}
-
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.April, 21, 12, 0, 0, 0, time.UTC)
-	nowISO := "2026-04-21T12:00:00.000Z"
-	project := storage.ProjectRecord{ID: "project_1", Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Projects.Upsert(ctx, project); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loop := storage.LoopRecord{ID: "loop_1", Seq: 30, ProjectID: project.ID, Type: "worker", TargetType: "project", Status: "running", CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Loops.Upsert(ctx, loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	lease, err := registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: loop.ID, RunID: "run_live", ExecutionID: "exec_live",
-	})
-	if err != nil {
-		t.Fatalf("AdmitSpawn() error = %v", err)
-	}
-	if lease.Context().Err() != nil {
-		t.Fatalf("lease already cancelled before close: %v", lease.Context().Err())
-	}
-
-	services := looperdruntime.Services{
-		Coordinator:      coordinator,
-		Repositories:     repos,
-		Loops:            &loops.Service{DB: coordinator.DB(), Repos: repos, Now: func() time.Time { return now }},
-		ActiveExecutions: registry,
-	}
-
-	// Force Runs.GetLatestByLoopID to fail after preflight would otherwise open
-	// the stop gate: close the SQLite coordinator under the services.
-	if err := coordinator.Close(); err != nil {
-		t.Fatalf("coordinator.Close() error = %v", err)
-	}
-
-	if _, err := closeLoop(ctx, services, loop.ID, "Closed by test", func() time.Time { return now }, nil, nil); err == nil {
-		t.Fatal("closeLoop() error = nil, want run lookup failure")
-	}
-	if err := lease.Context().Err(); err != nil {
-		t.Fatalf("lease cancelled after failed close preflight: %v (BeginLoopStop must run only after abortable lookups)", err)
-	}
-	if registry.LoopStopActive(loop.ID) {
-		t.Fatal("LoopStopActive = true after aborted close preflight, want gate never opened")
-	}
-	if _, err := registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: loop.ID, RunID: "run_after", ExecutionID: "exec_after",
-	}); err != nil {
-		t.Fatalf("AdmitSpawn after failed close preflight error = %v, want success", err)
-	}
-}
-
-// Failed Pause must not cancel spawn leases: BeginLoopStop is irreversible for
-// lease contexts wired into execution.run, so it runs only after Pause succeeds.
-func TestStopLoopDoesNotCancelLeasesWhenPauseFails(t *testing.T) {
-	ctx := context.Background()
-	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	if _, err := coordinator.MigrationRunner().RunPending(ctx); err != nil {
-		t.Fatalf("MigrationRunner().RunPending() error = %v", err)
-	}
-	t.Cleanup(func() { _ = coordinator.Close() })
-
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.April, 21, 12, 0, 0, 0, time.UTC)
-	nowISO := "2026-04-21T12:00:00.000Z"
-	project := storage.ProjectRecord{ID: "project_1", Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Projects.Upsert(ctx, project); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	// Terminal status cannot transition to paused → Pause fails while loop stays terminal.
-	loop := storage.LoopRecord{ID: "loop_1", Seq: 30, ProjectID: project.ID, Type: "worker", TargetType: "project", Status: "terminated", CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Loops.Upsert(ctx, loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	lease, err := registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: loop.ID, RunID: "run_pending", ExecutionID: "exec_pending",
-	})
-	if err != nil {
-		t.Fatalf("AdmitSpawn() error = %v", err)
-	}
-	if lease.Context().Err() != nil {
-		t.Fatalf("lease already cancelled before stop: %v", lease.Context().Err())
-	}
-
-	services := looperdruntime.Services{
-		Coordinator:      coordinator,
-		Repositories:     repos,
-		Loops:            &loops.Service{DB: coordinator.DB(), Repos: repos, Now: func() time.Time { return now }},
-		ActiveExecutions: registry,
-	}
-
-	if _, err := stopLoop(ctx, services, loop.ID, "Stopped by test", func() time.Time { return now }, nil, nil); err == nil {
-		t.Fatal("stopLoop() error = nil, want Pause transition failure")
-	}
-	if err := lease.Context().Err(); err != nil {
-		t.Fatalf("lease cancelled after failed Pause: %v (BeginLoopStop must run only after successful pause)", err)
-	}
-	// New spawns for the loop must still be admitted after the failed stop.
-	if _, err := registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: loop.ID, RunID: "run_after", ExecutionID: "exec_after",
-	}); err != nil {
-		t.Fatalf("AdmitSpawn after failed stop error = %v, want success (loop stop admission released)", err)
-	}
-}
-
 func TestCloseLoopDoesNotTerminateLoopWhenActiveKillFails(t *testing.T) {
 	ctx := context.Background()
 	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
@@ -772,210 +546,6 @@ func TestCloseLoopDoesNotTerminateLoopWhenActiveKillFails(t *testing.T) {
 		t.Fatalf("AdmitSpawn after aborted closeLoop error = %v, want success", err)
 	}
 }
-
-// When the Supervisor registry is present but has no entry for a stoppable
-// execution that still has a persisted PID, stop/close must fail loudly rather
-// than report success while leaving a live agent process behind (#576 / PR #586).
-func TestCloseLoopFailsWhenLiveHandleMissingWithPersistedPID(t *testing.T) {
-	ctx := context.Background()
-	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	if _, err := coordinator.MigrationRunner().RunPending(ctx); err != nil {
-		t.Fatalf("MigrationRunner().RunPending() error = %v", err)
-	}
-	t.Cleanup(func() { _ = coordinator.Close() })
-
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.April, 21, 12, 0, 0, 0, time.UTC)
-	nowISO := "2026-04-21T12:00:00.000Z"
-	project := storage.ProjectRecord{ID: "project_1", Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Projects.Upsert(ctx, project); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loop := storage.LoopRecord{ID: "loop_1", Seq: 30, ProjectID: project.ID, Type: "worker", TargetType: "project", Status: "running", CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Loops.Upsert(ctx, loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	run := storage.RunRecord{ID: "run_1", LoopID: loop.ID, Status: "running", StartedAt: nowISO, LastHeartbeatAt: &nowISO, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Runs.Upsert(ctx, run); err != nil {
-		t.Fatalf("Runs.Upsert() error = %v", err)
-	}
-	pid := int64(9876)
-	agentExecution := storage.AgentExecutionRecord{ID: "agentexec_1", ProjectID: &project.ID, LoopID: &loop.ID, RunID: &run.ID, Vendor: "codex", Status: "running", PID: &pid, StartedAt: nowISO, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.AgentExecutions.Upsert(ctx, agentExecution); err != nil {
-		t.Fatalf("AgentExecutions.Upsert() error = %v", err)
-	}
-
-	// Registry present but empty: no live handle for the running execution.
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	services := looperdruntime.Services{
-		Coordinator:      coordinator,
-		Repositories:     repos,
-		Loops:            &loops.Service{DB: coordinator.DB(), Repos: repos, Now: func() time.Time { return now }},
-		ActiveExecutions: registry,
-	}
-
-	signaled := false
-	if _, err := closeLoop(ctx, services, loop.ID, "Closed by test", func() time.Time { return now }, func(int, syscall.Signal) error {
-		signaled = true
-		return nil
-	}, nil); !errors.Is(err, looperdruntime.ErrAgentLiveHandleMissing) {
-		t.Fatalf("closeLoop() error = %v, want %v", err, looperdruntime.ErrAgentLiveHandleMissing)
-	}
-	if signaled {
-		t.Fatal("PID signal path invoked while Supervisor registry was present")
-	}
-	storedLoop, err := repos.Loops.GetByID(ctx, loop.ID)
-	if err != nil {
-		t.Fatalf("Loops.GetByID() error = %v", err)
-	}
-	if storedLoop == nil || storedLoop.Status != "running" {
-		t.Fatalf("Loops.GetByID() = %#v, want running loop after missing-handle failure", storedLoop)
-	}
-	if registry.LoopStopActive(loop.ID) {
-		t.Fatal("LoopStopActive = true after aborted closeLoop, want gate released")
-	}
-}
-
-func TestStopLoopFailsWhenLiveHandleMissingWithPersistedPID(t *testing.T) {
-	ctx := context.Background()
-	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	if _, err := coordinator.MigrationRunner().RunPending(ctx); err != nil {
-		t.Fatalf("MigrationRunner().RunPending() error = %v", err)
-	}
-	t.Cleanup(func() { _ = coordinator.Close() })
-
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.April, 21, 12, 0, 0, 0, time.UTC)
-	nowISO := "2026-04-21T12:00:00.000Z"
-	project := storage.ProjectRecord{ID: "project_1", Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Projects.Upsert(ctx, project); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loop := storage.LoopRecord{ID: "loop_1", Seq: 30, ProjectID: project.ID, Type: "worker", TargetType: "project", Status: "running", CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Loops.Upsert(ctx, loop); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	run := storage.RunRecord{ID: "run_1", LoopID: loop.ID, Status: "running", StartedAt: nowISO, LastHeartbeatAt: &nowISO, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.Runs.Upsert(ctx, run); err != nil {
-		t.Fatalf("Runs.Upsert() error = %v", err)
-	}
-	pid := int64(9876)
-	agentExecution := storage.AgentExecutionRecord{ID: "agentexec_1", ProjectID: &project.ID, LoopID: &loop.ID, RunID: &run.ID, Vendor: "codex", Status: "running", PID: &pid, StartedAt: nowISO, CreatedAt: nowISO, UpdatedAt: nowISO}
-	if err := repos.AgentExecutions.Upsert(ctx, agentExecution); err != nil {
-		t.Fatalf("AgentExecutions.Upsert() error = %v", err)
-	}
-
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	services := looperdruntime.Services{
-		Coordinator:      coordinator,
-		Repositories:     repos,
-		Loops:            &loops.Service{DB: coordinator.DB(), Repos: repos, Now: func() time.Time { return now }},
-		ActiveExecutions: registry,
-	}
-
-	if _, err := stopLoop(ctx, services, loop.ID, "Stopped by test", func() time.Time { return now }, func(int, syscall.Signal) error {
-		t.Fatal("PID signal path invoked while Supervisor registry was present")
-		return nil
-	}, nil); !errors.Is(err, looperdruntime.ErrAgentLiveHandleMissing) {
-		t.Fatalf("stopLoop() error = %v, want %v", err, looperdruntime.ErrAgentLiveHandleMissing)
-	}
-}
-
-// stopAll fallback after Pause failure must not leave a sticky loop stop gate:
-// stopCandidateExecution closes admission only for the kill window, then releases.
-func TestStopAllLoopsReleasesStopGateWhenPauseFails(t *testing.T) {
-	ctx := context.Background()
-	services, repos, now := newStopAllTestServices(t)
-	// Terminal loop status cannot transition to paused → stopLoop fails before
-	// durable pause, then stopAll falls back to stopCandidateExecution.
-	insertStopAllTestLoop(t, ctx, repos, now, stopAllLoopFixture{
-		loopID: "loop_failed_pause", seq: 1, loopType: "worker", loopStatus: "failed",
-		runID: "run_failed_pause", runStatus: "running",
-		executionID: "exec_failed_pause", executionStatus: "running", pid: 4201,
-	})
-
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	active := &fakeActiveExecution{}
-	unregister := registry.Register("loop_failed_pause", "run_failed_pause", "exec_failed_pause", active)
-	defer unregister()
-	services.ActiveExecutions = registry
-
-	response, err := stopAllLoops(ctx, services, "Stopped by test", func() time.Time { return now }, nil, nil)
-	if err != nil {
-		t.Fatalf("stopAllLoops() error = %v", err)
-	}
-	if response.Summary.Total != 1 {
-		t.Fatalf("stopAllLoops() summary = %#v, want Total=1", response.Summary)
-	}
-	// Pause fails (failed → paused is invalid); fallback still kills the live
-	// execution. Refresh may report alreadyStopping once status is cancelling.
-	if !active.killed {
-		t.Fatal("active execution Kill was not invoked on stop-all fallback")
-	}
-	storedLoop, err := repos.Loops.GetByID(ctx, "loop_failed_pause")
-	if err != nil {
-		t.Fatalf("Loops.GetByID() error = %v", err)
-	}
-	if storedLoop == nil || storedLoop.Status != "failed" {
-		t.Fatalf("Loops.GetByID() = %#v, want still-failed (not paused) loop", storedLoop)
-	}
-	if registry.LoopStopActive("loop_failed_pause") {
-		t.Fatal("LoopStopActive = true after stop-all Pause failure fallback, want gate released")
-	}
-	if _, err := registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: "loop_failed_pause", RunID: "run_after_fallback", ExecutionID: "exec_after_fallback",
-	}); err != nil {
-		t.Fatalf("AdmitSpawn after stop-all Pause failure fallback error = %v, want success", err)
-	}
-}
-
-// Successful stopAll must keep the sticky gate from haltLoop even though
-// stopCandidateExecution releases its nested BeginLoopStop refcount.
-func TestStopAllLoopsKeepsStopGateStickyAfterDurablePause(t *testing.T) {
-	ctx := context.Background()
-	services, repos, now := newStopAllTestServices(t)
-	insertStopAllTestLoop(t, ctx, repos, now, stopAllLoopFixture{
-		loopID: "loop_sticky", seq: 1, loopType: "worker", loopStatus: "running",
-		runID: "run_sticky", runStatus: "running",
-		executionID: "exec_sticky", executionStatus: "running", pid: 4202,
-	})
-
-	registry := looperdruntime.NewActiveExecutionRegistry()
-	active := &fakeActiveExecution{}
-	unregister := registry.Register("loop_sticky", "run_sticky", "exec_sticky", active)
-	defer unregister()
-	services.ActiveExecutions = registry
-
-	response, err := stopAllLoops(ctx, services, "Stopped by test", func() time.Time { return now }, nil, nil)
-	if err != nil {
-		t.Fatalf("stopAllLoops() error = %v", err)
-	}
-	if response.Summary.Failed != 0 {
-		t.Fatalf("stopAllLoops() summary = %#v, want no failures", response.Summary)
-	}
-	storedLoop, err := repos.Loops.GetByID(ctx, "loop_sticky")
-	if err != nil {
-		t.Fatalf("Loops.GetByID() error = %v", err)
-	}
-	if storedLoop == nil || storedLoop.Status != "paused" {
-		t.Fatalf("Loops.GetByID() = %#v, want paused after durable stop", storedLoop)
-	}
-	if !registry.LoopStopActive("loop_sticky") {
-		t.Fatal("LoopStopActive = false after durable stopAll, want sticky closed gate")
-	}
-	if _, err := registry.AdmitSpawn(ctx, agent.SpawnMeta{
-		LoopID: "loop_sticky", RunID: "run_blocked", ExecutionID: "exec_blocked",
-	}); !errors.Is(err, agent.ErrSpawnLoopStopping) {
-		t.Fatalf("AdmitSpawn after durable stopAll error = %v, want ErrSpawnLoopStopping", err)
-	}
-}
-
 func TestStopLoopKillsActiveInMemoryExecution(t *testing.T) {
 	ctx := context.Background()
 	coordinator, err := storage.OpenSQLiteCoordinator(ctx, filepath.Join(t.TempDir(), "looper.sqlite"), storage.SQLiteCoordinatorOptions{Migrations: storage.EmbeddedMigrations})
