+++ internal/api/loop_retry_discard_test.go
@@ -3,7 +3,6 @@ package api
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
@@ -15,7 +14,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/nexu-io/looper/internal/agent"
 	"github.com/nexu-io/looper/internal/domain"
 	gitinfra "github.com/nexu-io/looper/internal/infra/git"
 	looperdruntime "github.com/nexu-io/looper/internal/runtime"
@@ -565,149 +563,6 @@ func TestHandlerLoopRetryWithoutDiscardDoesNotReportDiscard(t *testing.T) {
 		t.Fatalf("worktreeDiscard present without flag: %#v", data["worktreeDiscard"])
 	}
 }
-
-// Retry must restore the sticky stop gate when a later TX validation fails so
-// a failed retry cannot reopen AdmitSpawn for stale pre-stop runners.
-func TestHandlerLoopRetryRestoresStopGateOnTXConflict(t *testing.T) {
-	rt, cfg := startTestRuntime(t)
-	h := NewHandler(Context{Config: cfg, Runtime: rt})
-	services := rt.Services()
-	nowISO := "2026-04-11T12:00:00.000Z"
-	projectID := "project_retry_restore_stop_gate"
-	loopID := "loop_retry_restore_stop_gate"
-	targetID := projectID
-	dedupeKey := "worker:restore_stop_gate"
-	otherLoopID := "loop_retry_restore_stop_gate_other"
-
-	if err := services.Repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{
-		ID: projectID, Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: loopID, Seq: 3141, ProjectID: projectID, Type: "worker", TargetType: "project",
-		TargetID: &targetID, Status: "paused", CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	lastErrorKind := "manual_intervention"
-	if err := services.Repositories.Queue.Upsert(context.Background(), storage.QueueItemRecord{
-		ID: "queue_retry_restore_stop_gate_failed", ProjectID: &projectID, LoopID: &loopID, Type: "worker",
-		TargetType: "project", TargetID: targetID, DedupeKey: dedupeKey,
-		Priority: storage.QueuePriorityWorker, Status: "failed", AvailableAt: nowISO,
-		Attempts: 1, MaxAttempts: 3, LastErrorKind: &lastErrorKind, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Queue.Upsert() failed item error = %v", err)
-	}
-	// Sibling loop exists but has no active queue yet so preflight dedupe passes.
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: otherLoopID, Seq: 3142, ProjectID: projectID, Type: "worker", TargetType: "project",
-		TargetID: &targetID, Status: "paused", CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() other loop error = %v", err)
-	}
-	if services.ActiveExecutions == nil {
-		t.Fatal("ActiveExecutions is nil")
-	}
-	if _, err := services.ActiveExecutions.BeginLoopStop(loopID, "looper stop"); err != nil {
-		t.Fatalf("BeginLoopStop: %v", err)
-	}
-	if !services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = false before retry, want sticky closed")
-	}
-
-	// After ClearLoopStop and before the requeue TX, inject an active dedupe
-	// row so the TX fails the way a concurrent requeue would.
-	h.retryAfterClearStopGateHook = func(id string) {
-		if id != loopID {
-			return
-		}
-		if services.ActiveExecutions.LoopStopActive(loopID) {
-			t.Error("LoopStopActive still true in after-clear hook, want gate already cleared")
-		}
-		if err := services.Repositories.Queue.Upsert(context.Background(), storage.QueueItemRecord{
-			ID: "queue_retry_restore_stop_gate_active", ProjectID: &projectID, LoopID: &otherLoopID, Type: "worker",
-			TargetType: "project", TargetID: targetID, DedupeKey: dedupeKey,
-			Priority: storage.QueuePriorityWorker, Status: "queued", AvailableAt: nowISO,
-			Attempts: 0, MaxAttempts: 3, CreatedAt: nowISO, UpdatedAt: nowISO,
-		}); err != nil {
-			t.Errorf("inject active dedupe: %v", err)
-		}
-	}
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/loops/3141/retry", strings.NewReader(`{"mode":"auto","resetAttempts":true}`))
-	recorder := httptest.NewRecorder()
-	h.ServeHTTP(recorder, req)
-	if recorder.Code != http.StatusConflict {
-		t.Fatalf("status = %d, want 409; body=%s", recorder.Code, recorder.Body.String())
-	}
-	if !services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = false after failed retry, want sticky gate restored")
-	}
-	if _, err := services.ActiveExecutions.AdmitSpawn(context.Background(), agent.SpawnMeta{
-		LoopID: loopID, RunID: "run_retry_restore", ExecutionID: "exec_retry_restore",
-	}); !errors.Is(err, agent.ErrSpawnLoopStopping) {
-		t.Fatalf("AdmitSpawn after failed retry error = %v, want ErrSpawnLoopStopping", err)
-	}
-}
-
-// Retry must clear the sticky stop gate before publishing claimable queue work
-// so a concurrent scheduler tick cannot claim then fail AdmitSpawn.
-func TestHandlerLoopRetryClearsStopGateBeforeClaimable(t *testing.T) {
-	rt, cfg := startTestRuntime(t)
-	h := NewHandler(Context{Config: cfg, Runtime: rt})
-	services := rt.Services()
-	nowISO := "2026-04-11T12:00:00.000Z"
-	projectID := "project_retry_clear_stop_gate"
-	loopID := "loop_retry_clear_stop_gate"
-	targetID := projectID
-
-	if err := services.Repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{
-		ID: projectID, Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: loopID, Seq: 3140, ProjectID: projectID, Type: "worker", TargetType: "project",
-		TargetID: &targetID, Status: "paused", CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	lastErrorKind := "manual_intervention"
-	if err := services.Repositories.Queue.Upsert(context.Background(), storage.QueueItemRecord{
-		ID: "queue_retry_clear_stop_gate", ProjectID: &projectID, LoopID: &loopID, Type: "worker",
-		TargetType: "project", TargetID: targetID, DedupeKey: "worker:clear_stop_gate",
-		Priority: storage.QueuePriorityWorker, Status: "failed", AvailableAt: nowISO,
-		Attempts: 1, MaxAttempts: 3, LastErrorKind: &lastErrorKind, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Queue.Upsert() error = %v", err)
-	}
-	if services.ActiveExecutions == nil {
-		t.Fatal("ActiveExecutions is nil")
-	}
-	if _, err := services.ActiveExecutions.BeginLoopStop(loopID, "looper stop"); err != nil {
-		t.Fatalf("BeginLoopStop: %v", err)
-	}
-	if !services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = false before retry, want sticky closed")
-	}
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/loops/3140/retry", strings.NewReader(`{"mode":"auto","resetAttempts":true}`))
-	recorder := httptest.NewRecorder()
-	h.ServeHTTP(recorder, req)
-	if recorder.Code != http.StatusOK {
-		t.Fatalf("status = %d, want 200; body=%s", recorder.Code, recorder.Body.String())
-	}
-	if services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = true after retry, want ClearLoopStop before claimable queue publish")
-	}
-	if _, err := services.ActiveExecutions.AdmitSpawn(context.Background(), agent.SpawnMeta{
-		LoopID: loopID, RunID: "run_retry_clear", ExecutionID: "exec_retry_clear",
-	}); err != nil {
-		t.Fatalf("AdmitSpawn after retry error = %v, want success", err)
-	}
-}
-
 func TestHandlerLoopRetryDiscardPreservesDirtyWorktreeWhenAgentNotConfigured(t *testing.T) {
 	rt, cfg := startTestRuntime(t)
 	cfg.Agent.Vendor = nil
@@ -907,236 +762,6 @@ func TestHandlerLoopRetryDiscardAllowsCompletedSiblingPRLoop(t *testing.T) {
 		t.Fatalf("dirty.txt still present after discard with completed sibling: %v", err)
 	}
 }
-
-// TestHandlerLoopStartClearsStopGateBeforeClaimable ensures start/unpause clears
-// the sticky stop gate before the requeue TX commits claimable work (same race
-// as retry: concurrent tick must not see running+queued with gate still closed).
-func TestHandlerLoopStartClearsStopGateBeforeClaimable(t *testing.T) {
-	rt, cfg := startTestRuntime(t)
-	h := NewHandler(Context{Config: cfg, Runtime: rt})
-	services := rt.Services()
-	nowISO := "2026-04-11T12:00:00.000Z"
-	projectID := "project_start_clear_stop_gate"
-	loopID := "loop_start_clear_stop_gate"
-	targetID := projectID
-	metadata := `{"worker":{"title":"Start gate","prompt":"go","repo":"acme/looper","baseBranch":"main"}}`
-
-	if err := services.Repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{
-		ID: projectID, Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: loopID, Seq: 3150, ProjectID: projectID, Type: "worker", TargetType: "project",
-		TargetID: &targetID, Status: "paused", MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	if services.ActiveExecutions == nil {
-		t.Fatal("ActiveExecutions is nil")
-	}
-	if _, err := services.ActiveExecutions.BeginLoopStop(loopID, "looper stop"); err != nil {
-		t.Fatalf("BeginLoopStop: %v", err)
-	}
-	if !services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = false before start, want sticky closed")
-	}
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/loops/"+loopID+"/start", nil)
-	recorder := httptest.NewRecorder()
-	h.ServeHTTP(recorder, req)
-	if recorder.Code != http.StatusOK {
-		t.Fatalf("start status = %d, want 200; body=%s", recorder.Code, recorder.Body.String())
-	}
-	if services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = true after start, want ClearLoopStop before claimable queue publish")
-	}
-	if _, err := services.ActiveExecutions.AdmitSpawn(context.Background(), agent.SpawnMeta{
-		LoopID: loopID, RunID: "run_start_clear", ExecutionID: "exec_start_clear",
-	}); err != nil {
-		t.Fatalf("AdmitSpawn after start error = %v, want success", err)
-	}
-}
-
-// TestHandlerLoopStartRestoresStopGateOnTXFailure ensures a failed start does
-// not leave AdmitSpawn open after a sticky looper-stop gate was pre-cleared.
-func TestHandlerLoopStartRestoresStopGateOnTXFailure(t *testing.T) {
-	rt, cfg := startTestRuntime(t)
-	h := NewHandler(Context{Config: cfg, Runtime: rt})
-	services := rt.Services()
-	nowISO := "2026-04-11T12:00:00.000Z"
-	projectID := "project_start_restore_stop_gate"
-	loopID := "loop_start_restore_stop_gate"
-	otherLoopID := "loop_start_restore_other"
-	targetID := "pr:acme/looper:77"
-	repo := "acme/looper"
-	prNumber := int64(77)
-	metadata := `{"worker":{"title":"Start restore","prompt":"go","repo":"acme/looper","baseBranch":"main"}}`
-
-	if err := services.Repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{
-		ID: projectID, Name: "Looper", RepoPath: t.TempDir(), CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: loopID, Seq: 3151, ProjectID: projectID, Type: "worker", TargetType: "pull_request",
-		TargetID: &targetID, Repo: &repo, PRNumber: &prNumber, Status: "paused",
-		MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() loop error = %v", err)
-	}
-	// Conflicting active loop for the same PR target so start TX fails unique check.
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: otherLoopID, Seq: 3152, ProjectID: projectID, Type: "worker", TargetType: "pull_request",
-		TargetID: &targetID, Repo: &repo, PRNumber: &prNumber, Status: "running",
-		MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() other error = %v", err)
-	}
-	if services.ActiveExecutions == nil {
-		t.Fatal("ActiveExecutions is nil")
-	}
-	if _, err := services.ActiveExecutions.BeginLoopStop(loopID, "looper stop"); err != nil {
-		t.Fatalf("BeginLoopStop: %v", err)
-	}
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/loops/"+loopID+"/start", nil)
-	recorder := httptest.NewRecorder()
-	h.ServeHTTP(recorder, req)
-	if recorder.Code == http.StatusOK {
-		t.Fatalf("start status = 200, want conflict/error; body=%s", recorder.Body.String())
-	}
-	if !services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = false after failed start, want sticky gate restored")
-	}
-	if _, err := services.ActiveExecutions.AdmitSpawn(context.Background(), agent.SpawnMeta{
-		LoopID: loopID, RunID: "run_start_restore", ExecutionID: "exec_start_restore",
-	}); !errors.Is(err, agent.ErrSpawnLoopStopping) {
-		t.Fatalf("AdmitSpawn after failed start error = %v, want ErrSpawnLoopStopping", err)
-	}
-}
-
-// TestHandlerWorkersCreateReuseClearsStickyStopGate ensures issue-worker reuse
-// (paused → queued) reopens the sticky stop spawn gate closed by looper stop.
-// Without this, recreate-same-issue claims the queue then AgentExecutor.Start
-// fails with ErrSpawnLoopStopping forever.
-func TestHandlerWorkersCreateReuseClearsStickyStopGate(t *testing.T) {
-	rt, cfg := startTestRuntime(t)
-	h := NewHandler(Context{Config: cfg, Runtime: rt})
-	services := rt.Services()
-	nowISO := "2026-04-11T12:00:00.000Z"
-
-	baseBranch := "main"
-	metadata := `{"repo":"acme/looper"}`
-	if err := services.Repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{
-		ID: "project_worker_reuse_stop_gate", Name: "Looper", RepoPath: "/tmp/repos/looper",
-		BaseBranch: &baseBranch, MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-
-	loopID := "loop_worker_reuse_stop_gate"
-	targetID := "issue:acme/looper:99"
-	repo := "acme/looper"
-	workerMeta := `{"worker":{"title":"Stopped issue worker","repo":"acme/looper","baseBranch":"main","issueNumber":99}}`
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: loopID, Seq: 3125, ProjectID: "project_worker_reuse_stop_gate", Type: "worker",
-		TargetType: "issue", TargetID: &targetID, Repo: &repo, Status: "paused",
-		MetadataJSON: &workerMeta, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	if services.ActiveExecutions == nil {
-		t.Fatal("ActiveExecutions is nil")
-	}
-	// Simulate sticky gate left by successful looper stop (pause without release).
-	if _, err := services.ActiveExecutions.BeginLoopStop(loopID, "looper stop"); err != nil {
-		t.Fatalf("BeginLoopStop: %v", err)
-	}
-	if !services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = false before reuse, want sticky closed")
-	}
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/workers", strings.NewReader(
-		`{"projectId":"project_worker_reuse_stop_gate","repo":"acme/looper","issueNumber":99,"baseBranch":"main"}`,
-	))
-	req.Header.Set("content-type", "application/json")
-	recorder := httptest.NewRecorder()
-	h.ServeHTTP(recorder, req)
-	if recorder.Code != http.StatusOK {
-		t.Fatalf("worker reuse status = %d body=%s, want 200", recorder.Code, recorder.Body.String())
-	}
-	if services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = true after issue-worker reuse, want ClearLoopStop for queued reactivation")
-	}
-	if _, err := services.ActiveExecutions.AdmitSpawn(context.Background(), agent.SpawnMeta{
-		LoopID: loopID, RunID: "run_reuse", ExecutionID: "exec_reuse",
-	}); err != nil {
-		t.Fatalf("AdmitSpawn after worker reuse error = %v, want success", err)
-	}
-}
-
-// TestHandlerWorkersCreateReuseRestoresStopGateOnTXFailure ensures a failed
-// issue-worker reuse restores the sticky stop gate when it was pre-cleared
-// before the reuse TX published claimable work. force+running is still found as
-// a conflicting issue worker (pre-clear runs) then resume fails with conflict.
-func TestHandlerWorkersCreateReuseRestoresStopGateOnTXFailure(t *testing.T) {
-	rt, cfg := startTestRuntime(t)
-	h := NewHandler(Context{Config: cfg, Runtime: rt})
-	services := rt.Services()
-	nowISO := "2026-04-11T12:00:00.000Z"
-
-	baseBranch := "main"
-	metadata := `{"repo":"acme/looper"}`
-	if err := services.Repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{
-		ID: "project_worker_reuse_restore_gate", Name: "Looper", RepoPath: "/tmp/repos/looper",
-		BaseBranch: &baseBranch, MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-
-	loopID := "loop_worker_reuse_restore_gate"
-	targetID := "issue:acme/looper:98"
-	repo := "acme/looper"
-	workerMeta := `{"worker":{"title":"Running issue worker","repo":"acme/looper","baseBranch":"main","issueNumber":98}}`
-	if err := services.Repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: loopID, Seq: 3126, ProjectID: "project_worker_reuse_restore_gate", Type: "worker",
-		TargetType: "issue", TargetID: &targetID, Repo: &repo, Status: "running",
-		MetadataJSON: &workerMeta, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	if services.ActiveExecutions == nil {
-		t.Fatal("ActiveExecutions is nil")
-	}
-	if _, err := services.ActiveExecutions.BeginLoopStop(loopID, "looper stop"); err != nil {
-		t.Fatalf("BeginLoopStop: %v", err)
-	}
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/workers", strings.NewReader(
-		`{"projectId":"project_worker_reuse_restore_gate","repo":"acme/looper","issueNumber":98,"baseBranch":"main","force":true}`,
-	))
-	req.Header.Set("content-type", "application/json")
-	recorder := httptest.NewRecorder()
-	h.ServeHTTP(recorder, req)
-	if recorder.Code == http.StatusOK {
-		t.Fatalf("worker reuse status = 200, want error; body=%s", recorder.Body.String())
-	}
-	if !services.ActiveExecutions.LoopStopActive(loopID) {
-		t.Fatal("LoopStopActive = false after failed force reuse of running worker, want sticky gate restored")
-	}
-	if _, err := services.ActiveExecutions.AdmitSpawn(context.Background(), agent.SpawnMeta{
-		LoopID: loopID, RunID: "run_reuse_restore", ExecutionID: "exec_reuse_restore",
-	}); !errors.Is(err, agent.ErrSpawnLoopStopping) {
-		t.Fatalf("AdmitSpawn after failed reuse error = %v, want ErrSpawnLoopStopping", err)
-	}
-}
-
-// TestHandlerWorkersCreateReuseSharesRetryLockWithDiscard ensures POST /workers
-// issue-worker reuse takes the same per-loop mutex as discard+retry, so reuse
-// cannot enqueue between discard preflight and git reset (wiping the worktree
-// for the reuse-created queue item, then failing retry with active-queue conflict).
 func TestHandlerWorkersCreateReuseSharesRetryLockWithDiscard(t *testing.T) {
 	rt, cfg := startTestRuntime(t)
 	h := NewHandler(Context{Config: cfg, Runtime: rt})
